+++ gems/aws-sdk-dax/lib/aws-sdk-dax/client.rb
@@ -1420,7 +1420,7 @@ module Aws::DAX
         params: params,
         config: config)
       context[:gem_name] = 'aws-sdk-dax'
-      context[:gem_version] = '1.0.0'
+      context[:gem_version] = '1.0.1'
       Seahorse::Client::Request.new(handlers, context)
     end