+++ gems/aws-sdk-marketplacemetering/lib/aws-sdk-marketplacemetering/client.rb
@@ -323,7 +323,7 @@ module Aws::MarketplaceMetering
         params: params,
         config: config)
       context[:gem_name] = 'aws-sdk-marketplacemetering'
-      context[:gem_version] = '1.0.0'
+      context[:gem_version] = '1.0.1'
       Seahorse::Client::Request.new(handlers, context)
     end