+++ gems/aws-sdk-resourcegroupstaggingapi/lib/aws-sdk-resourcegroupstaggingapi/client.rb
@@ -485,7 +485,7 @@ module Aws::ResourceGroupsTaggingAPI
         params: params,
         config: config)
       context[:gem_name] = 'aws-sdk-resourcegroupstaggingapi'
-      context[:gem_version] = '1.0.0'
+      context[:gem_version] = '1.0.1'
       Seahorse::Client::Request.new(handlers, context)
     end