+++ gems/aws-sdk-swf/lib/aws-sdk-swf/client.rb
@@ -3819,7 +3819,7 @@ module Aws::SWF
         params: params,
         config: config)
       context[:gem_name] = 'aws-sdk-swf'
-      context[:gem_version] = '1.0.0'
+      context[:gem_version] = '1.0.1'
       Seahorse::Client::Request.new(handlers, context)
     end